vdlimit is a replacement for the prior tools used to set disklimits.
The prior tools appear to have been a hack on top of the quota support.

NOTE that the vdlimit code was taken from Herbert Poetzl's Experimental
website. It is vdlimit 0.01. The code didn't compile and was standalone.
I integrated it with util-vserver, but it needs further integration. Have
already sent Herbert email about this and another minor problem.