Add the Emulation Component support to the PLCAPI.
The Emulation Component is a part of the work
done by UniPi in the context of the EU OneLab project.

The purpose of this component is to emulate
the behaviour of a wired or a wireless link.
It is based on an external device, called
Dummynet box, sitting on the link between Nodes and
the rest of the Internet.

The PLCAPI extension involves:
 - modification to the planetlab database
   to store DummynetBoxes;
 - a python `DummyBoxes' class and Methods to manipulate
   these additional information.

In order to generate a new Dummynet box image, an additional
package is needed, containing the picobsd.bin base image.

The roles check is still in development.

More information regarding the main architectural
of the Emulation Component could be found
on the 2nd and 3rd deliverables(*) provided 
as part of the OneLab project.

(*)
http://www.one-lab.org/pub/OneLab/PublicDeliverables/OneLab-deliverable-D4E-2-Integrated-dummynet-and-PlanetLab.pdf
the 3rd will be available soon.