A disabled or diagnostic-mode node will contact PLC (download BM and interact with the API), but there is no indication of this via the API.
This modification updates the 'last_contact' field just like GetSlivers.

Specifically, this allows the Node&PCU Registration Wizard to set a node to
'disable', and then wait for a tech to boot it and for the node to contact
PLC.  The wizard now has an indication, without entering production mode, that
the node has booted, and therefore that the wizard can continue.

This not fool-proof, since if the tech stops here, the node will remain in the
DB, but it will be 'disabled'.