Needed to change the semantics of SetSlice(), which previously would delete *all* tags associated with a slice.  This would wipe out sliver tags that are set at runtime by both NM and other scripts.
SetSlice() now ignores sliver tags.  The assumption is that the global
defaults that are being set by db-config are just that: global
defaults.  If someone want to override those defaults or introduce new
tags then one should do so using sliver tags.

This is motivated by the hmac tag that is set at run time by NM.  But
there are other tags for which we want similar semantics.