Introduced method mapping so that XMLRPC calls can be forwarded to alternate methods.
This was needed to support the GENI AM because one of the GENI calls is Resolve, which means that there would be a methods/Resolve.py and a methods/resolve.py, which causes problems on filesystems (like HFS+) that don't support caps.  So, now the GENI AM simply creates a mapping from Resolve to ResolveGENI.  

This could also be useful if you wanted to host try new methods without replacing the old ones first.